Finally pointed the DEET in the right direction.
g_image is the unlensed image of the source.  g_lensimage is the lensed
image of the source.  So, sample_vis.uvmodel was operating on the
unlensed source image, not the lensed source image.  No wonder it kept
failing to find a good fit!  This probably explains the difficulty I’ve
had modeling FLS3 as well.